requirements.txt
pyyaml
requests
websocket-client
python-dotenv
# 假设需要的库，具体以LongPort SDK要求为准
longport